+++ assume/markets/clearing_algorithms/all_or_nothing.py
@@ -199,4 +199,7 @@ class PayAsBidAonRole(MarketRole):
                     product,
                 )
             )
-        return accepted_orders, rejected_orders, meta
+            # wirte network flows here if applicable
+            flows = []
+
+        return accepted_orders, rejected_orders, meta, flows