+++ assume/markets/clearing_algorithms/contracts.py
@@ -264,8 +264,11 @@ class PayAsBidContractRole(MarketRole):
                     partial(self.execute_contract, contract=order), recurrency_task
                 )
 
+            #write flows if applicable
+            flows=[]
+
         # contract clearing (pay_as_bid) takes place
-        return accepted_orders, rejected_orders, meta
+        return accepted_orders, rejected_orders, meta, flows
 
     async def execute_contract(self, contract: Order):
         """